refactor: move verify tuple function closer to usage
